requirements.txt
asgiref==3.5.2
boto3==1.20.0
botocore==1.23.45
Brotli==1.0.9
cement==2.8.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==3.2.9
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-storages==1.13.1
djangorestframework==3.14.0
future==0.16.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.1
oauthlib==3.2.2
pathspec==0.9.0
postgres==4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pycparser==2.21
PyJWT==2.6.0
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
pywin32==305
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
s3transfer==0.5.2
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.3
style==1.1.0
termcolor==1.1.0
tzdata==2022.7
update==0.0.1
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.1.9
whitenoise==6.2.0
